+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.4.3
+ - Animation Frames are no longer be sent after disconnect
+ - Animation Frames are no longer sent if new device frame data is not available
+ - suppressAnimationLoop, when set to false in node, will cause node to try and use the animation loop, and fail.
+
 ## 0.4.2 - 2014-03-12
  - Allow controller.use to be called idempotently
  - Fix issue where requestAnimationFrame would not be used in Chrome